test cases updated, UI and code refactoring.
