+++ flexmeasures/ui/templates/base.html
@@ -935,7 +935,7 @@
                         <h5 class="card-title">${kpi.title}</h5>
                         <h5 class="card-text">
                             <strong>
-                                ${label}: ${kpi.downsample_value}
+                                ${label}: ${kpi.downsample_value} ${kpi.unit}
                                 <small>${kpi.downsample_function}</small>
                             </strong>
                         </h5>
