+++ flexmeasures/utils/coding_utils.py
@@ -177,27 +177,3 @@ def find_classes_modules(module, superclass, skiptest=True):
 
 def get_classes_module(module, superclass, skiptest=True) -> dict:
     return dict(find_classes_modules(module, superclass, skiptest=skiptest))
-
-
-from flexmeasures.data.models.time_series import Sensor  # noqa E402
-
-
-def get_downsample_function_and_value(
-    kpi: dict, sensor: Sensor, sensor_stats: dict
-) -> tuple:
-    downsample_function = kpi.get("default_function", None)
-    if downsample_function is None:
-        if sensor.unit == "%":
-            downsample_function = "mean"
-        else:
-            downsample_function = "sum"
-    try:
-        if downsample_function == "mean":
-            downsample_value = dict(next(iter(sensor_stats.values())))["Mean value"]
-        else:
-            downsample_value = dict(next(iter(sensor_stats.values())))[
-                "Sum over values"
-            ]
-    except StopIteration:
-        downsample_value = 0
-    return downsample_function, downsample_value
